+++ test-cases/syntax-1-instanced-fn/src3/index.js
@@ -1,1 +1,0 @@
-console.log(JSON.stringify({ ...require('../src1/styles_1.scss'), ...require('../src1/styles_2.scss') }, null, 2));