+++ test-cases/syntax-2-static-fn-with-plugin/webpack.build1.config.js
@@ -1,1 +1,1 @@
-module.exports = require('./webpack.config.maker.js')(1, 'create-map');
+module.exports = require('./webpack.config.maker.js')(1);