+++ test-cases/syntax-2-static-fn-with-plugin/webpack.build2.config.js
@@ -1,1 +1,1 @@
-module.exports = require('./webpack.config.maker.js')(2, 'extend-map');
+module.exports = require('./webpack.config.maker.js')(2);